Webapp, more client and integrator improvements

+++ src/main/java/pl/amitec/mercury/clients/bitbee/types/Tax.java
@@ -5,12 +5,14 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record Tax (
-    @JsonProperty Integer id,
+    @JsonProperty Long id,
     @JsonProperty String name,
-    @JsonProperty int percent,
-    @JsonProperty("shop") Integer shopId
+    @JsonProperty BigDecimal percent,
+    @JsonProperty("shop") Long shopId
 ) {}
